backend/requirements.txt
fastapi>=0.136.0
uvicorn[standard]>=0.44.0
pydantic>=2.13.0
pydantic-settings>=2.13.0
sqlalchemy>=2.0.49
pgvector>=0.4.2
httpx>=0.28.1
python-multipart>=0.0.26
anthropic>=0.49.0
openai>=1.70.0
python-dotenv>=1.2.0
numpy>=2.4.0
